pull encode and decode out into its own class
