+++ Gemfile
@@ -2,4 +2,5 @@ source "https://rubygems.org"
 
 gem "minitest"
 gem "pry-state"
+gem 'pry'
