+++ test/night_writer_test.rb
@@ -7,13 +7,6 @@ class NightWriterTest < Minitest::Test
   def writer
     @writer ||= NightWriter.new
   end
-  def test_can_lookup
-    assert_equal ".", writer.lookup("a", 4)
-  end
-
-  def test_postion
-    assert_equal ".", writer.lookup(:capitalize, 0)
-  end
 
   def test_can_encode_a_letter
     assert_equal "0.\n..\n..\n", writer.encode_to_braille("a")
